Better isolation of UCI extensions
