+++ README.md
@@ -21,6 +21,8 @@ try (UCI uci = new UCI(engine)) {
 }
 ```
 
+If you plan to use the *perft* and *test* commands described below, launch the **ExtendedUCI** class instead of **UCI**.
+
 That's all!
 
 ## Partial implementation ... and extra functionalities
@@ -29,19 +31,22 @@ It does not directly support the following commands (but you can add them in a *
 - **register**: As a promoter of open source free software, I won't encourage you to develop software that requires registration.
 - **ponderhit** has not yet been implemented.
 
-It also does not recognize commands starting with unknown token (to be honest, it's not very hard to implement but seems a very bad, error prone, idea to me).
-
-It implements the following extensions:
-- **q** is a shortcut for standard **quit** command.
+It also does not recognize commands starting with unknown token (to be honest, it's not very hard to implement but seemed a very bad, error prone, idea to me).  
+The following extensions are implemented in **UCI** class:
 - It can accept different engines, that can be selected using the **engine** command. You can view these engines as plugins.  
-**engine** [*engineId*] lists the available engines' ids or changes the engine if *engineId* is provided.
-- **d** [*fen*] displays a textual representation of the game. If the command is followed by *fen*, the command displays the representation of a game in the [Forsyth–Edwards Notation](https://en.wikipedia.org/wiki/Forsyth%E2%80%93Edwards_Notation).</li>
+- **engine** [*engineId*] lists the available engines' ids or changes the engine if *engineId* is provided.
+- **q** is a shortcut for standard **quit** command
+
+The **ExtendedUCI** class implements the following extensions in addition of the standard commands:
+- **d** [*fen*] displays a textual representation of the game. If the command is followed by *fen*, the command displays the representation of a game in the [Forsyth–Edwards Notation](https://en.wikipedia.org/wiki/Forsyth%E2%80%93Edwards_Notation).  
+**Please note this command is optional**, only engines that implement *com.fathzer.jchess.uci.extended.Displayable* interface support it.
 - **perft** *depth* [threads *nb*] [legal] [playleaves] runs [perft](https://www.chessprogramming.org/Perft) test and displays the divide result.  
 *depth* is mandatory and is the search depth of the perft algorithm. It should be strictly positive.  
 *threads* is followed by the number of threads used to process the queries. This number should be strictly positive. Default is 1. *threads* can be replaced by the shortcut *t*.  
 *legal* uses legal moves from the move generator instead of the default pseudo-legal moves. *legal* can be replaced by the shortcut *l*.  
 *playleaves* plays, when used with *legal*, the leave moves. These moves are always played when using pseudo-legal moves. *playleaves* can be replaced by the shortcut *pl*.     
-**Please note this command is optional**, only engines that implement *com.fathzer.jchess.uci.MoveGeneratorSupplier* interface support it.
+**Please note this command is optional**, only engines that implement *com.fathzer.jchess.uci.extended.MoveGeneratorSupplier* interface support it.  
+If the engine implements the **com.fathzer.jchess.uci.extended.MoveToUCIConverter** interface, the divide displays moves in UCI format, otherwise the move's toString method is used.
 - **test** *depth* [threads *nb*] [legal] [playleaves] [cut *s*] runs a move generator test based on [perft](https://www.chessprogramming.org/Perft).  
 It can also be used to test move generator's performance as it outputs the number of moves generated per second.  
 *depth* is mandatory and is the search depth of the perft algorithm. It should be strictly positive.  
@@ -62,10 +67,9 @@ protected Collection<PerfTTestData> readTestData() {
 	}
 }
 ``` 
-- **q** is a shortcut for standard **quit** command</li>
 
 ## Adding custom commands
-Override the **com.fathzer.jchess.uci.UCI** class and use its *addCommand* method to add your own custom commands.  
+Override the **com.fathzer.jchess.uci.UCI** or ****com.fathzer.jchess.uci.extended.ExtendedUCI** classes and use its *addCommand* method to add your own custom commands.  
 Then instantiate your UCI subclass and launch its **run** method.
 
 ## Get rid of System.out and System.in
@@ -74,6 +78,5 @@ If you want another way to exchange messages, you can subclass the UCI class and
 
 
 ## TODO
-* Make games-core optional. It would implies to split UCI into two classes, one with the strict UCI command set, and a subclass that adds PerfT related things.
 * Verify the engine is protected against strange client behavior (like changing the position during a go request).
 * Implement support for pondering.
