--- src/main/java/com/fathzer/jchess/uci/MoveToUCIConverter.java
@@ -1,5 +1,0 @@
-package com.fathzer.jchess.uci;
-
-public interface MoveToUCIConverter<M> {
-	UCIMove toUCI(M move);
-}
